+++ example/lib/src/home/view/page_02.dart
@@ -16,23 +16,23 @@ class Page2 extends StatefulWidget {
   @override
   State createState() => _Page2State();
 
-  void onPressed() {
-    //
-    final state = con.ofState<_Page2State>()!;
-
-    state.setState(() {
-      con.incrementCounter();
-    });
-  }
+  /// This will 'rebuild' the InheritedWidget defined below
+  /// and not call setState() function to instead initiate a rebuild.
+  void onPressed() => con.incrementCounter();
 }
 
 /// This works with a separate 'data source'
 /// It doesn't no what, but being so, the count is never reset to zero.
-class _Page2State extends StateMVC<Page2> {
+class _Page2State extends InheritedStateMVC<Page2, _Page02Inherited> {
+  /// Define an InheritedWidget to be inserted above this Widget on the Widget tree.
+  _Page2State()
+      : super(inheritedWidget: (child) => _Page02Inherited(child: child));
+
   @override
   void initState() {
     /// Make this the 'current' State object for the Controller.
     add(widget.con);
+
     con = controller as Controller;
 
     /// Allow for con.initState() function to fire.
@@ -42,54 +42,70 @@ class _Page2State extends StateMVC<Page2> {
   //
   late Controller con;
 
+  /// Define the 'child' Widget that will be passed to the InheritedWidget above.
   @override
-  Widget build(BuildContext context) => BuildPage(
-        label: '2',
-        count: con.count,
-        counter: widget.onPressed,
-        row: (context) => [
-          Flexible(
-            child: ElevatedButton(
-              key: const Key('Page 1'),
-              style: flatButtonStyle,
-              onPressed: () {
-                Navigator.pop(context);
-              },
-              child: const Text(
-                'Page 1',
-              ),
-            ),
-          ),
-          Flexible(
-            child: ElevatedButton(
-              key: const Key('Page 3'),
-              style: flatButtonStyle,
-              onPressed: () async {
-                await Navigator.push(
-                    context,
-                    MaterialPageRoute<void>(
-                        builder: (BuildContext context) => Page3()));
+  Widget buildChild(BuildContext context) {
+    /// Comment this command out and the counter will not work.
+    /// That's because this Widget is then no longer a dependency to the InheritedWidget above.
+    widgetInherited(context);
 
-                /// A good habit to get into. Refresh the screen again.
-                /// In this case, to show the count may have changed.
-                setState(() {});
-              },
-              child: const Text(
-                'Page 3',
-              ),
+    /// Ignore BuildPage(). It's used only to highlight the other features in this page
+    return BuildPage(
+      label: '2',
+      count: con.count,
+      counter: widget.onPressed,
+      row: (context) => [
+        Flexible(
+          child: ElevatedButton(
+            key: const Key('Page 1'),
+            style: flatButtonStyle,
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            child: const Text(
+              'Page 1',
             ),
           ),
-        ],
-        column: (context) => [
-          const Text("Has a 'data source' to save the count"),
-        ],
-        persistentFooterButtons: <Widget>[
-          ElevatedButton(
-            key: const Key('Page 1 Counter'),
+        ),
+        Flexible(
+          child: ElevatedButton(
+            key: const Key('Page 3'),
             style: flatButtonStyle,
-            onPressed: Page1().onPressed,
-            child: const Text('Page 1 Counter'),
+            onPressed: () async {
+              await Navigator.push(
+                  context,
+                  MaterialPageRoute<void>(
+                      builder: (BuildContext context) => Page3()));
+
+              /// A good habit to get into. Refresh the screen again.
+              /// In this case, to show the count may have changed.
+              setState(() {});
+            },
+            child: const Text(
+              'Page 3',
+            ),
           ),
-        ],
-      );
+        ),
+      ],
+      column: (context) => [
+        const Text("Has a 'data source' to save the count"),
+      ],
+      persistentFooterButtons: <Widget>[
+        ElevatedButton(
+          key: const Key('Page 1 Counter'),
+          style: flatButtonStyle,
+          onPressed: Page1().onPressed,
+          child: const Text('Page 1 Counter'),
+        ),
+      ],
+    );
+  }
+}
+
+/// The inserted InheritedWidget that takes in the buildChild() Widget above.
+class _Page02Inherited extends InheritedWidget {
+  const _Page02Inherited({Key? key, required Widget child})
+      : super(key: key, child: child);
+  @override
+  bool updateShouldNotify(oldWidget) => true;
 }